@@ -1,5 +1,4 @@
-import Link from "next/link";
-import Image from "next/image";
+import Link from 'next/link';
 
 export default function Footer() {
   return (
@@ -9,47 +8,22 @@ export default function Footer() {
           {/* Brand */}
           <div className="col-span-1">
             <Link href="/" className="block mb-4">
-              <span className="font-heading text-h2 font-bold text-white">
-                HandcraftBK
-              </span>
+              <span className="font-heading text-h2 font-bold text-white">HandcraftBK</span>
             </Link>
-            <p className="text-subtle mb-6">
-              HandcraftBK made with love and passion by artisans around the
-              world.
-            </p>
+            <p className="text-subtle mb-6">HandcraftBK made with love and passion by artisans around the world.</p>
             <div className="flex space-x-4">
-              <a
-                href="https://www.facebook.com/chuquanpho29"
-                target="_blank"
-                aria-label="Facebook"
-              >
-                <svg
-                  className="w-6 h-6 text-white"
-                  fill="currentColor"
-                  viewBox="0 0 24 24"
-                >
+              <a href="https://www.facebook.com/chuquanpho29" target="_blank" aria-label="Facebook">
+                <svg className="w-6 h-6 text-white" fill="currentColor" viewBox="0 0 24 24">
                   <path d="M22 12c0-5.523-4.477-10-10-10S2 6.477 2 12c0 4.991 3.657 9.128 8.438 9.878v-6.987h-2.54V12h2.54V9.797c0-2.506 1.492-3.89 3.777-3.89 1.094 0 2.238.195 2.238.195v2.46h-1.26c-1.243 0-1.63.771-1.63 1.562V12h2.773l-.443 2.89h-2.33v6.988C18.343 21.128 22 16.991 22 12z"></path>
                 </svg>
               </a>
-              <a
-                href="https://www.instagram.com/hunglt29"
-                target="_blank"
-                aria-label="Instagram"
-              >
-                <svg
-                  className="w-6 h-6 text-white"
-                  fill="currentColor"
-                  viewBox="0 0 24 24"
-                >
+              <a href="https://www.instagram.com/hunglt29" target="_blank" aria-label="Instagram">
+                <svg className="w-6 h-6 text-white" fill="currentColor" viewBox="0 0 24 24">
                   <path d="M12.315 2c2.43 0 2.784.013 3.808.06 1.064.049 1.791.218 2.427.465a4.902 4.902 0 011.772 1.153 4.902 4.902 0 011.153 1.772c.247.636.416 1.363.465 2.427.048 1.067.06 1.407.06 4.123v.08c0 2.643-.012 2.987-.06 4.043-.049 1.064-.218 1.791-.465 2.427a4.902 4.902 0 01-1.153 1.772 4.902 4.902 0 01-1.772 1.153c-.636.247-1.363.416-2.427.465-1.067.048-1.407.06-4.123.06h-.08c-2.643 0-2.987-.012-4.043-.06-1.064-.049-1.791-.218-2.427-.465a4.902 4.902 0 01-1.772-1.153 4.902 4.902 0 01-1.153-1.772c-.247-.636-.416-1.363-.465-2.427-.047-1.024-.06-1.379-.06-3.808v-.63c0-2.43.013-2.784.06-3.808.049-1.064.218-1.791.465-2.427a4.902 4.902 0 011.153-1.772A4.902 4.902 0 015.45 2.525c.636-.247 1.363-.416 2.427-.465C8.901 2.013 9.256 2 11.685 2h.63zm-.081 1.802h-.468c-2.456 0-2.784.011-3.807.058-.975.045-1.504.207-1.857.344-.467.182-.8.398-1.15.748-.35.35-.566.683-.748 1.15-.137.353-.3.882-.344 1.857-.047 1.023-.058 1.351-.058 3.807v.468c0 2.456.011 2.784.058 3.807.045.975.207 1.504.344 1.857.182.466.399.8.748 1.15.35.35.683.566 1.15.748.353.137.882.3 1.857.344 1.054.048 1.37.058 4.041.058h.08c2.597 0 2.917-.01 3.96-.058.976-.045 1.505-.207 1.858-.344.466-.182.8-.398 1.15-.748.35-.35.566-.683.748-1.15.137-.353.3-.882.344-1.857.048-1.055.058-1.37.058-4.041v-.08c0-2.597-.01-2.917-.058-3.96-.045-.976-.207-1.505-.344-1.858a3.097 3.097 0 00-.748-1.15 3.098 3.098 0 00-1.15-.748c-.353-.137-.882-.3-1.857-.344-1.023-.047-1.351-.058-3.807-.058zM12 6.865a5.135 5.135 0 110 10.27 5.135 5.135 0 010-10.27zm0 1.802a3.333 3.333 0 100 6.666 3.333 3.333 0 000-6.666zm5.338-3.205a1.2 1.2 0 110 2.4 1.2 1.2 0 010-2.4z"></path>
                 </svg>
               </a>
               <a href="#" target="_blank" aria-label="Pinterest">
-                <svg
-                  className="w-6 h-6 text-white"
-                  fill="currentColor"
-                  viewBox="0 0 24 24"
-                >
+                <svg className="w-6 h-6 text-white" fill="currentColor" viewBox="0 0 24 24">
                   <path d="M12.017 0C5.396 0 .029 5.367.029 11.987c0 5.079 3.158 9.417 7.618 11.162-.105-.949-.199-2.403.041-3.439.219-.937 1.406-5.957 1.406-5.957s-.359-.72-.359-1.781c0-1.663.967-2.911 2.168-2.911 1.024 0 1.518.769 1.518 1.688 0 1.029-.653 2.567-.992 3.992-.285 1.193.6 2.165 1.775 2.165 2.128 0 3.768-2.245 3.768-5.487 0-2.861-2.063-4.869-5.008-4.869-3.41 0-5.409 2.562-5.409 5.199 0 1.033.394 2.143.889 2.741.099.12.112.225.085.345-.09.375-.293 1.199-.334 1.363-.053.225-.172.271-.4.165-1.495-.69-2.433-2.878-2.433-4.646 0-3.776 2.748-7.252 7.92-7.252 4.158 0 7.392 2.967 7.392 6.923 0 4.135-2.607 7.462-6.233 7.462-1.214 0-2.354-.629-2.758-1.379l-.749 2.848c-.269 1.045-1.004 2.352-1.498 3.146 1.123.345 2.306.535 3.55.535 6.607 0 11.985-5.365 11.985-11.987C23.97 5.39 18.592.026 11.985.026L12.017 0z"></path>
                 </svg>
               </a>
@@ -93,42 +67,27 @@ export default function Footer() {
             <h3 className="text-white text-lg font-bold mb-4">Categories</h3>
             <ul className="space-y-2">
               <li>
-                <Link
-                  href="/category/home-decor"
-                  className="text-subtle hover:text-white"
-                >
+                <Link href="/category/home-decor" className="text-subtle hover:text-white">
                   Home Decor
                 </Link>
               </li>
               <li>
-                <Link
-                  href="/category/kitchen"
-                  className="text-subtle hover:text-white"
-                >
+                <Link href="/category/kitchen" className="text-subtle hover:text-white">
                   Kitchen
                 </Link>
               </li>
               <li>
-                <Link
-                  href="/category/wall-art"
-                  className="text-subtle hover:text-white"
-                >
+                <Link href="/category/wall-art" className="text-subtle hover:text-white">
                   Wall Art
                 </Link>
               </li>
               <li>
-                <Link
-                  href="/category/textiles"
-                  className="text-subtle hover:text-white"
-                >
+                <Link href="/category/textiles" className="text-subtle hover:text-white">
                   Textiles
                 </Link>
               </li>
               <li>
-                <Link
-                  href="/category/ceramics"
-                  className="text-subtle hover:text-white"
-                >
+                <Link href="/category/ceramics" className="text-subtle hover:text-white">
                   Ceramics
                 </Link>
               </li>
@@ -140,18 +99,12 @@ export default function Footer() {
             <h3 className="text-white text-lg font-bold mb-4">Account</h3>
             <ul className="space-y-2">
               <li>
-                <Link
-                  href="/account/profile"
-                  className="text-subtle hover:text-white"
-                >
+                <Link href="/account/profile" className="text-subtle hover:text-white">
                   My Profile
                 </Link>
               </li>
               <li>
-                <Link
-                  href="/account/purchase"
-                  className="text-subtle hover:text-white"
-                >
+                <Link href="/account/purchase" className="text-subtle hover:text-white">
                   My purchases
                 </Link>
               </li>
@@ -160,9 +113,7 @@ export default function Footer() {
         </div>
 
         <div className="border-t border-accent mt-12 pt-6 text-center">
-          <p className="text-subtle text-small">
-            © {new Date().getFullYear()} HandcraftBK. All rights reserved.
-          </p>
+          <p className="text-subtle text-small">© {new Date().getFullYear()} HandcraftBK. All rights reserved.</p>
         </div>
       </div>
     </footer>
