@@ -28,7 +28,7 @@ const CartContext = createContext<CartContextType | undefined>(undefined);
 
 export function CartProvider({ children }: { children: ReactNode }) {
   const [items, setItems] = useState<CartItem[]>([]);
-  
+
   // Load cart from localStorage on mount
   useEffect(() => {
     const savedCart = localStorage.getItem('cart');
@@ -41,22 +41,20 @@ export function CartProvider({ children }: { children: ReactNode }) {
       }
     }
   }, []);
-  
+
   // Save cart to localStorage when it changes
   useEffect(() => {
     localStorage.setItem('cart', JSON.stringify(items));
   }, [items]);
-  
+
   const addToCart = (product: Product, quantity: number) => {
     setItems(prevItems => {
       const existingItem = prevItems.find(item => item.product.id === product.id);
-      
+
       if (existingItem) {
         // Update quantity if product already in cart
-        return prevItems.map(item => 
-          item.product.id === product.id 
-            ? { ...item, quantity: item.quantity + quantity } 
-            : item
+        return prevItems.map(item =>
+          item.product.id === product.id ? { ...item, quantity: item.quantity + quantity } : item
         );
       } else {
         // Add new item to cart
@@ -64,44 +62,40 @@ export function CartProvider({ children }: { children: ReactNode }) {
       }
     });
   };
-  
+
   const removeFromCart = (productId: number) => {
     setItems(prevItems => prevItems.filter(item => item.product.id !== productId));
   };
-  
+
   const updateQuantity = (productId: number, quantity: number) => {
     if (quantity <= 0) {
       removeFromCart(productId);
       return;
     }
-    
-    setItems(prevItems => 
-      prevItems.map(item => 
-        item.product.id === productId 
-          ? { ...item, quantity } 
-          : item
-      )
-    );
+
+    setItems(prevItems => prevItems.map(item => (item.product.id === productId ? { ...item, quantity } : item)));
   };
-  
+
   const clearCart = () => {
     setItems([]);
   };
-  
+
   const totalItems = items.reduce((total, item) => total + item.quantity, 0);
-  
-  const subtotal = items.reduce((total, item) => total + (item.product.price * item.quantity), 0);
-  
+
+  const subtotal = items.reduce((total, item) => total + item.product.price * item.quantity, 0);
+
   return (
-    <CartContext.Provider value={{
-      items,
-      addToCart,
-      removeFromCart,
-      updateQuantity,
-      clearCart,
-      totalItems,
-      subtotal
-    }}>
+    <CartContext.Provider
+      value={{
+        items,
+        addToCart,
+        removeFromCart,
+        updateQuantity,
+        clearCart,
+        totalItems,
+        subtotal,
+      }}
+    >
       {children}
     </CartContext.Provider>
   );
