@@ -1,28 +1,24 @@
-import { useState } from "react";
-import Image from "next/image";
-import Layout from "../components/Layout";
+import { useState } from 'react';
+import Image from 'next/image';
+import Layout from '../components/Layout';
 
 export default function Contact() {
   const [formState, setFormState] = useState({
-    name: "",
-    email: "",
-    subject: "",
-    message: "",
+    name: '',
+    email: '',
+    subject: '',
+    message: '',
   });
 
   const [formStatus, setFormStatus] = useState({
     submitted: false,
     success: false,
-    message: "",
+    message: '',
   });
 
-  const handleChange = (
-    e: React.ChangeEvent<
-      HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement
-    >
-  ) => {
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement>) => {
     const { name, value } = e.target;
-    setFormState((prev) => ({ ...prev, [name]: value }));
+    setFormState(prev => ({ ...prev, [name]: value }));
   };
 
   const handleSubmit = (e: React.FormEvent) => {
@@ -33,7 +29,7 @@ export default function Contact() {
       setFormStatus({
         submitted: true,
         success: false,
-        message: "Please fill in all required fields.",
+        message: 'Please fill in all required fields.',
       });
       return;
     }
@@ -43,13 +39,13 @@ export default function Contact() {
       setFormStatus({
         submitted: true,
         success: true,
-        message: "Thank you for your message! We will get back to you soon.",
+        message: 'Thank you for your message! We will get back to you soon.',
       });
       setFormState({
-        name: "",
-        email: "",
-        subject: "",
-        message: "",
+        name: '',
+        email: '',
+        subject: '',
+        message: '',
       });
     }, 1000);
   };
@@ -71,9 +67,7 @@ export default function Contact() {
         </div>
         <div className="relative container mx-auto px-4 h-full flex items-center">
           <div>
-            <h1 className="text-white font-heading text-4xl md:text-5xl font-bold mb-2">
-              Contact Us
-            </h1>
+            <h1 className="text-white font-heading text-4xl md:text-5xl font-bold mb-2">Contact Us</h1>
             <p className="text-white text-lg">We'd love to hear from you</p>
           </div>
         </div>
@@ -89,9 +83,7 @@ export default function Contact() {
               {formStatus.submitted && (
                 <div
                   className={`p-4 mb-6 rounded ${
-                    formStatus.success
-                      ? "bg-green-100 text-green-700"
-                      : "bg-red-100 text-red-700"
+                    formStatus.success ? 'bg-green-100 text-green-700' : 'bg-red-100 text-red-700'
                   }`}
                 >
                   {formStatus.message}
@@ -138,7 +130,7 @@ export default function Contact() {
                       tabIndex={0}
                       className="flex justify-between items-center w-full px-3 py-3 border border-gray-300 rounded cursor-pointer hover:border-primary focus:outline-none focus:ring-1 focus:ring-primary"
                     >
-                      {formState.subject || "Please select"}
+                      {formState.subject || 'Please select'}
                       <svg
                         xmlns="http://www.w3.org/2000/svg"
                         className="h-5 w-5 ml-2"
@@ -146,12 +138,7 @@ export default function Contact() {
                         viewBox="0 0 24 24"
                         stroke="currentColor"
                       >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M19 9l-7 7-7-7"
-                        />
+                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
                       </svg>
                     </label>
                     <ul
@@ -161,9 +148,9 @@ export default function Contact() {
                       <li>
                         <button
                           onClick={() => {
-                            setFormState((prev) => ({
+                            setFormState(prev => ({
                               ...prev,
-                              subject: "Product Inquiry",
+                              subject: 'Product Inquiry',
                             }));
                           }}
                           className="py-2 px-4 hover:bg-gray-100 text-left"
@@ -174,9 +161,9 @@ export default function Contact() {
                       <li>
                         <button
                           onClick={() => {
-                            setFormState((prev) => ({
+                            setFormState(prev => ({
                               ...prev,
-                              subject: "Order Status",
+                              subject: 'Order Status',
                             }));
                           }}
                           className="py-2 px-4 hover:bg-gray-100 text-left"
@@ -187,9 +174,9 @@ export default function Contact() {
                       <li>
                         <button
                           onClick={() => {
-                            setFormState((prev) => ({
+                            setFormState(prev => ({
                               ...prev,
-                              subject: "Return or Exchange",
+                              subject: 'Return or Exchange',
                             }));
                           }}
                           className="py-2 px-4 hover:bg-gray-100 text-left"
@@ -200,9 +187,9 @@ export default function Contact() {
                       <li>
                         <button
                           onClick={() => {
-                            setFormState((prev) => ({
+                            setFormState(prev => ({
                               ...prev,
-                              subject: "Wholesale Inquiry",
+                              subject: 'Wholesale Inquiry',
                             }));
                           }}
                           className="py-2 px-4 hover:bg-gray-100 text-left"
@@ -213,9 +200,9 @@ export default function Contact() {
                       <li>
                         <button
                           onClick={() => {
-                            setFormState((prev) => ({
+                            setFormState(prev => ({
                               ...prev,
-                              subject: "Artisan Partnership",
+                              subject: 'Artisan Partnership',
                             }));
                           }}
                           className="py-2 px-4 hover:bg-gray-100 text-left"
@@ -226,9 +213,9 @@ export default function Contact() {
                       <li>
                         <button
                           onClick={() => {
-                            setFormState((prev) => ({
+                            setFormState(prev => ({
                               ...prev,
-                              subject: "Other",
+                              subject: 'Other',
                             }));
                           }}
                           className="py-2 px-4 hover:bg-gray-100 text-left"
@@ -319,29 +306,20 @@ export default function Contact() {
                   </div>
                   <div className="ml-9">
                     <p className="text-gray-600 mb-1">
-                      General Inquiries:{" "}
-                      <a
-                        href="mailto:chuquanpho29@gmail.com"
-                        className="text-primary hover:underline"
-                      >
+                      General Inquiries:{' '}
+                      <a href="mailto:chuquanpho29@gmail.com" className="text-primary hover:underline">
                         chuquanpho29@gmail.com
                       </a>
                     </p>
                     <p className="text-gray-600 mb-1">
-                      Customer Support:{" "}
-                      <a
-                        href="mailto:chuquanpho29@gmail.com"
-                        className="text-primary hover:underline"
-                      >
+                      Customer Support:{' '}
+                      <a href="mailto:chuquanpho29@gmail.com" className="text-primary hover:underline">
                         chuquanpho29@gmail.com
                       </a>
                     </p>
                     <p className="text-gray-600">
-                      Artisan Relations:{" "}
-                      <a
-                        href="mailto:chuquanpho29@gmail.com"
-                        className="text-primary hover:underline"
-                      >
+                      Artisan Relations:{' '}
+                      <a href="mailto:chuquanpho29@gmail.com" className="text-primary hover:underline">
                         chuquanpho29@gmail.com
                       </a>
                     </p>
@@ -368,17 +346,12 @@ export default function Contact() {
                   </div>
                   <div className="ml-9">
                     <p className="text-gray-600 mb-1">
-                      Phone:{" "}
-                      <a
-                        href="tel:+84783225944"
-                        className="text-primary hover:underline"
-                      >
+                      Phone:{' '}
+                      <a href="tel:+84783225944" className="text-primary hover:underline">
                         84783225944
                       </a>
                     </p>
-                    <p className="text-gray-600">
-                      Hours: Monday-Friday, 9am-6pm PT
-                    </p>
+                    <p className="text-gray-600">Hours: Monday-Friday, 9am-6pm PT</p>
                   </div>
                 </div>
 
@@ -402,29 +375,17 @@ export default function Contact() {
                   </div>
                   <div className="ml-9 flex space-x-4">
                     <a href="#" className="text-primary hover:text-accent">
-                      <svg
-                        className="w-6 h-6"
-                        fill="currentColor"
-                        viewBox="0 0 24 24"
-                      >
+                      <svg className="w-6 h-6" fill="currentColor" viewBox="0 0 24 24">
                         <path d="M22 12c0-5.523-4.477-10-10-10S2 6.477 2 12c0 4.991 3.657 9.128 8.438 9.878v-6.987h-2.54V12h2.54V9.797c0-2.506 1.492-3.89 3.777-3.89 1.094 0 2.238.195 2.238.195v2.46h-1.26c-1.243 0-1.63.771-1.63 1.562V12h2.773l-.443 2.89h-2.33v6.988C18.343 21.128 22 16.991 22 12z"></path>
                       </svg>
                     </a>
                     <a href="#" className="text-primary hover:text-accent">
-                      <svg
-                        className="w-6 h-6"
-                        fill="currentColor"
-                        viewBox="0 0 24 24"
-                      >
+                      <svg className="w-6 h-6" fill="currentColor" viewBox="0 0 24 24">
                         <path d="M12.315 2c2.43 0 2.784.013 3.808.06 1.064.049 1.791.218 2.427.465a4.902 4.902 0 011.772 1.153 4.902 4.902 0 011.153 1.772c.247.636.416 1.363.465 2.427.048 1.067.06 1.407.06 4.123v.08c0 2.643-.012 2.987-.06 4.043-.049 1.064-.218 1.791-.465 2.427a4.902 4.902 0 01-1.153 1.772 4.902 4.902 0 01-1.772 1.153c-.636.247-1.363.416-2.427.465-1.067.048-1.407.06-4.123.06h-.08c-2.643 0-2.987-.012-4.043-.06-1.064-.049-1.791-.218-2.427-.465a4.902 4.902 0 01-1.772-1.153 4.902 4.902 0 01-1.153-1.772c-.247-.636-.416-1.363-.465-2.427-.047-1.024-.06-1.379-.06-3.808v-.63c0-2.43.013-2.784.06-3.808.049-1.064.218-1.791.465-2.427a4.902 4.902 0 011.153-1.772A4.902 4.902 0 015.45 2.525c.636-.247 1.363-.416 2.427-.465C8.901 2.013 9.256 2 11.685 2h.63zm-.081 1.802h-.468c-2.456 0-2.784.011-3.807.058-.975.045-1.504.207-1.857.344-.467.182-.8.398-1.15.748-.35.35-.566.683-.748 1.15-.137.353-.3.882-.344 1.857-.047 1.023-.058 1.351-.058 3.807v.468c0 2.456.011 2.784.058 3.807.045.975.207 1.504.344 1.857.182.466.399.8.748 1.15.35.35.683.566 1.15.748.353.137.882.3 1.857.344 1.054.048 1.37.058 4.041.058h.08c2.597 0 2.917-.01 3.96-.058.976-.045 1.505-.207 1.858-.344.466-.182.8-.398 1.15-.748.35-.35.566-.683.748-1.15.137-.353.3-.882.344-1.857.048-1.055.058-1.37.058-4.041v-.08c0-2.597-.01-2.917-.058-3.96-.045-.976-.207-1.505-.344-1.858a3.097 3.097 0 00-.748-1.15 3.098 3.098 0 00-1.15-.748c-.353-.137-.882-.3-1.857-.344-1.023-.047-1.351-.058-3.807-.058zM12 6.865a5.135 5.135 0 110 10.27 5.135 5.135 0 010-10.27zm0 1.802a3.333 3.333 0 100 6.666 3.333 3.333 0 000-6.666zm5.338-3.205a1.2 1.2 0 110 2.4 1.2 1.2 0 010-2.4z"></path>
                       </svg>
                     </a>
                     <a href="#" className="text-primary hover:text-accent">
-                      <svg
-                        className="w-6 h-6"
-                        fill="currentColor"
-                        viewBox="0 0 24 24"
-                      >
+                      <svg className="w-6 h-6" fill="currentColor" viewBox="0 0 24 24">
                         <path d="M12.017 0C5.396 0 .029 5.367.029 11.987c0 5.079 3.158 9.417 7.618 11.162-.105-.949-.199-2.403.041-3.439.219-.937 1.406-5.957 1.406-5.957s-.359-.72-.359-1.781c0-1.663.967-2.911 2.168-2.911 1.024 0 1.518.769 1.518 1.688 0 1.029-.653 2.567-.992 3.992-.285 1.193.6 2.165 1.775 2.165 2.128 0 3.768-2.245 3.768-5.487 0-2.861-2.063-4.869-5.008-4.869-3.41 0-5.409 2.562-5.409 5.199 0 1.033.394 2.143.889 2.741.099.12.112.225.085.345-.09.375-.293 1.199-.334 1.363-.053.225-.172.271-.4.165-1.495-.69-2.433-2.878-2.433-4.646 0-3.776 2.748-7.252 7.92-7.252 4.158 0 7.392 2.967 7.392 6.923 0 4.135-2.607 7.462-6.233 7.462-1.214 0-2.354-.629-2.758-1.379l-.749 2.848c-.269 1.045-1.004 2.352-1.498 3.146 1.123.345 2.306.535 3.55.535 6.607 0 11.985-5.365 11.985-11.987C23.97 5.39 18.592.026 11.985.026L12.017 0z"></path>
                       </svg>
                     </a>
@@ -466,52 +427,38 @@ export default function Contact() {
       {/* FAQ Section */}
       <section className="py-16 bg-background">
         <div className="container mx-auto px-4">
-          <h2 className="font-heading text-h1 text-center mb-12">
-            Frequently Asked Questions
-          </h2>
+          <h2 className="font-heading text-h1 text-center mb-12">Frequently Asked Questions</h2>
 
           <div className="grid grid-cols-1 md:grid-cols-2 gap-6 max-w-4xl mx-auto">
             <div className="bg-white p-6 rounded-lg shadow-sm">
-              <h3 className="font-heading text-h3 mb-3">
-                What are your shipping times?
-              </h3>
+              <h3 className="font-heading text-h3 mb-3">What are your shipping times?</h3>
               <p className="text-gray-600">
-                We typically process orders within 1-2 business days. Domestic
-                shipping takes 3-5 business days, while international shipping
-                can take 7-14 business days depending on the destination.
+                We typically process orders within 1-2 business days. Domestic shipping takes 3-5 business days, while
+                international shipping can take 7-14 business days depending on the destination.
               </p>
             </div>
 
             <div className="bg-white p-6 rounded-lg shadow-sm">
-              <h3 className="font-heading text-h3 mb-3">
-                Do you offer returns or exchanges?
-              </h3>
+              <h3 className="font-heading text-h3 mb-3">Do you offer returns or exchanges?</h3>
               <p className="text-gray-600">
-                Yes, we offer returns and exchanges within 30 days of purchase.
-                Items must be in original condition. Please note that custom
-                orders are non-returnable.
+                Yes, we offer returns and exchanges within 30 days of purchase. Items must be in original condition.
+                Please note that custom orders are non-returnable.
               </p>
             </div>
 
             <div className="bg-white p-6 rounded-lg shadow-sm">
-              <h3 className="font-heading text-h3 mb-3">
-                How do I track my order?
-              </h3>
+              <h3 className="font-heading text-h3 mb-3">How do I track my order?</h3>
               <p className="text-gray-600">
-                Once your order ships, you'll receive a confirmation email with
-                tracking information. You can also log into your account to view
-                order status.
+                Once your order ships, you'll receive a confirmation email with tracking information. You can also log
+                into your account to view order status.
               </p>
             </div>
 
             <div className="bg-white p-6 rounded-lg shadow-sm">
-              <h3 className="font-heading text-h3 mb-3">
-                Can I request custom orders?
-              </h3>
+              <h3 className="font-heading text-h3 mb-3">Can I request custom orders?</h3>
               <p className="text-gray-600">
-                Absolutely! We love working on custom pieces. Please contact us
-                with your ideas and requirements, and we'll connect you with an
-                artisan who specializes in that craft.
+                Absolutely! We love working on custom pieces. Please contact us with your ideas and requirements, and
+                we'll connect you with an artisan who specializes in that craft.
               </p>
             </div>
           </div>
@@ -530,8 +477,7 @@ export default function Contact() {
         <div className="container mx-auto px-4 text-center">
           <h2 className="font-heading text-h1 mb-4">Stay Updated</h2>
           <p className="max-w-2xl mx-auto mb-8">
-            Subscribe to our newsletter for updates on new artisans,
-            collections, and exclusive offers.
+            Subscribe to our newsletter for updates on new artisans, collections, and exclusive offers.
           </p>
 
           <form className="max-w-md mx-auto flex flex-col sm:flex-row">
