@@ -1,10 +1,10 @@
-import { useState, useEffect, useRef } from "react";
-import Image from "next/image";
-import Link from "next/link";
-import Layout from "../components/Layout";
-import ProductCard from "../components/ProductCard";
-import CategoryCard from "../components/CategoryCard";
-import TestimonialCard from "../components/TestimonialCard";
+import { useState, useEffect, useRef } from 'react';
+import Image from 'next/image';
+import Link from 'next/link';
+import Layout from '../components/Layout';
+import ProductCard from '../components/ProductCard';
+import CategoryCard from '../components/CategoryCard';
+import TestimonialCard from '../components/TestimonialCard';
 
 type Product = {
   id: number;
@@ -24,51 +24,44 @@ export default function Home() {
   const carouselRef = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
-    fetch("/api/hello")
-      .then((response) => response.json())
-      .then((data) => {
-        const featured = data.products.filter(
-          (product: Product) => product.featured
-        );
-        console.log("Featured Products:", featured);
+    fetch('/api/hello')
+      .then(response => response.json())
+      .then(data => {
+        const featured = data.products.filter((product: Product) => product.featured);
         setFeaturedProducts(featured);
       });
   }, []);
 
   const nextSlide = () => {
     if (featuredProducts.length <= 3) return;
-    setCurrentIndex((prevIndex) =>
-      prevIndex + 3 >= featuredProducts.length ? 0 : prevIndex + 3
-    );
+    setCurrentIndex(prevIndex => (prevIndex + 3 >= featuredProducts.length ? 0 : prevIndex + 3));
   };
 
   const prevSlide = () => {
     if (featuredProducts.length <= 3) return;
-    setCurrentIndex((prevIndex) =>
-      prevIndex - 3 < 0
-        ? Math.floor((featuredProducts.length - 1) / 3) * 3
-        : prevIndex - 3
+    setCurrentIndex(prevIndex =>
+      prevIndex - 3 < 0 ? Math.floor((featuredProducts.length - 1) / 3) * 3 : prevIndex - 3
     );
   };
 
   const categories = [
     {
-      name: "Home Decor",
-      image: "/images/insta-1.jpg",
+      name: 'Home Decor',
+      image: '/images/insta-1.jpg',
       productCount: 24,
-      slug: "home-decor",
+      slug: 'home-decor',
     },
     {
-      name: "Kitchen",
-      image: "/images/insta-1.jpg",
+      name: 'Kitchen',
+      image: '/images/insta-1.jpg',
       productCount: 18,
-      slug: "kitchen",
+      slug: 'kitchen',
     },
     {
-      name: "Wall Art",
-      image: "/images/insta-1.jpg",
+      name: 'Wall Art',
+      image: '/images/insta-1.jpg',
       productCount: 12,
-      slug: "wall-art",
+      slug: 'wall-art',
     },
   ];
 
@@ -76,20 +69,20 @@ export default function Home() {
     {
       quote:
         "The ceramic vase I purchased is absolutely stunning. The craftsmanship is impeccable and it's become the centerpiece of my living room.",
-      customerName: "Sarah Johnson",
-      location: "New York, NY",
+      customerName: 'Sarah Johnson',
+      location: 'New York, NY',
     },
     {
       quote:
         "I'm impressed by the quality of the handwoven baskets. They're not only beautiful but incredibly durable. Will definitely buy more!",
-      customerName: "Michael Thompson",
-      location: "Portland, OR",
+      customerName: 'Michael Thompson',
+      location: 'Portland, OR',
     },
     {
       quote:
-        "The macramé wall hanging has transformed my bedroom. The attention to detail is remarkable. So happy with my purchase!",
-      customerName: "Emily Rodriguez",
-      location: "Austin, TX",
+        'The macramé wall hanging has transformed my bedroom. The attention to detail is remarkable. So happy with my purchase!',
+      customerName: 'Emily Rodriguez',
+      location: 'Austin, TX',
     },
   ];
 
@@ -110,12 +103,9 @@ export default function Home() {
         </div>
         <div className="relative container mx-auto px-4 h-full flex items-center">
           <div className="max-w-xl">
-            <h1 className="text-white text-4xl md:text-5xl font-bold mb-4">
-              Welcome to HandcraftBK!
-            </h1>
+            <h1 className="text-white text-4xl md:text-5xl font-bold mb-4">Welcome to HandcraftBK!</h1>
             <p className="text-white text-lg mb-8">
-              Discover unique handmade items created with passion and skill by
-              artisans from around the world.
+              Discover unique handmade items created with passion and skill by artisans from around the world.
             </p>
             <Link href="/products" className="btn-primary inline-block">
               Explore Our Collection
@@ -127,9 +117,7 @@ export default function Home() {
       {/* Featured Categories */}
       <section className="py-16 bg-background">
         <div className="container mx-auto px-4">
-          <h2 className="font-heading text-center text-h1 mb-12">
-            Featured Categories
-          </h2>
+          <h2 className="font-heading text-center text-h1 mb-12">Featured Categories</h2>
           <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
             {categories.map((category, index) => (
               <CategoryCard
@@ -152,9 +140,7 @@ export default function Home() {
       {/* Best Selling Products */}
       <section className="py-10">
         <div className="container mx-auto px-4">
-          <h2 className="font-heading text-center text-h1 mb-12">
-            Best Selling Products
-          </h2>
+          <h2 className="font-heading text-center text-h1 mb-12">Best Selling Products</h2>
           {featuredProducts.length > 0 ? (
             <div className="relative px-8">
               {/* Carousel navigation buttons */}
@@ -170,12 +156,7 @@ export default function Home() {
                   viewBox="0 0 24 24"
                   stroke="currentColor"
                 >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M15 19l-7-7 7-7"
-                  />
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" />
                 </svg>
               </button>
 
@@ -191,12 +172,7 @@ export default function Home() {
                   viewBox="0 0 24 24"
                   stroke="currentColor"
                 >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M9 5l7 7-7 7"
-                  />
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
                 </svg>
               </button>
 
@@ -206,7 +182,7 @@ export default function Home() {
                   className="flex transition-transform duration-500 ease-in-out"
                   style={{ transform: `translateX(-${currentIndex * 33.33}%)` }}
                 >
-                  {featuredProducts.map((product) => (
+                  {featuredProducts.map(product => (
                     <div
                       key={product.id}
                       className="w-full sm:w-1/2 lg:w-1/3 flex-shrink-0 flex items-center justify-center"
@@ -236,9 +212,7 @@ export default function Home() {
                     key={index}
                     onClick={() => setCurrentIndex(index * 3)}
                     className={`w-2 h-2 rounded-full ${
-                      Math.floor(currentIndex / 3) === index
-                        ? "bg-primary"
-                        : "bg-gray-300"
+                      Math.floor(currentIndex / 3) === index ? 'bg-primary' : 'bg-gray-300'
                     }`}
                     aria-label={`Go to slide ${index + 1}`}
                   />
@@ -263,12 +237,10 @@ export default function Home() {
           <div className="flex items-center justify-center flex-col">
             <h2 className="font-heading text-h1 mb-6">Our Story</h2>
             <p className="mb-4 text-center">
-              We started HandcraftBK with a simple mission: to celebrate and
-              share the beauty of handmade crafts with the world while
-              supporting skilled artisans. Each item in our collection is
-              carefully selected for its quality, uniqueness, and the story
-              behind it. We work directly with artisans to ensure fair
-              compensation and sustainable practices.
+              We started HandcraftBK with a simple mission: to celebrate and share the beauty of handmade crafts with
+              the world while supporting skilled artisans. Each item in our collection is carefully selected for its
+              quality, uniqueness, and the story behind it. We work directly with artisans to ensure fair compensation
+              and sustainable practices.
             </p>
             <Link href="/about" className="btn-text">
               Learn More About Us
@@ -280,9 +252,7 @@ export default function Home() {
       {/* Testimonials */}
       <section className="py-10">
         <div className="container mx-auto px-4">
-          <h2 className="font-heading text-center text-h1 mb-12">
-            What Our Customers Say
-          </h2>
+          <h2 className="font-heading text-center text-h1 mb-12">What Our Customers Say</h2>
           <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
             {testimonials.map((testimonial, index) => (
               <TestimonialCard
@@ -302,11 +272,9 @@ export default function Home() {
           <div className="grid grid-cols-1 lg:grid-cols-2 gap-12">
             {/* Instagram Feed */}
             <div>
-              <h2 className="font-heading text-h2 mb-6">
-                Follow Us on Instagram
-              </h2>
+              <h2 className="font-heading text-h2 mb-6">Follow Us on Instagram</h2>
               <div className="grid grid-cols-3 gap-2">
-                {[1, 2, 3].map((item) => (
+                {[1, 2, 3].map(item => (
                   <div key={item} className="relative aspect-square">
                     <Image
                       src={`/images/insta-${item}.jpg`}
@@ -330,13 +298,8 @@ export default function Home() {
 
             {/* Newsletter */}
             <div className="bg-white p-8 rounded-lg">
-              <h2 className="font-heading text-h2 mb-4">
-                Subscribe to Our Newsletter
-              </h2>
-              <p className="mb-6">
-                Be the first to know about new arrivals, special offers, and
-                artisan stories.
-              </p>
+              <h2 className="font-heading text-h2 mb-4">Subscribe to Our Newsletter</h2>
+              <p className="mb-6">Be the first to know about new arrivals, special offers, and artisan stories.</p>
               <form className="flex flex-col sm:flex-row">
                 <input
                   type="email"
@@ -352,8 +315,7 @@ export default function Home() {
                 </button>
               </form>
               <p className="text-small text-gray-500 mt-4">
-                By subscribing, you agree to our Privacy Policy and consent to
-                receive updates from us.
+                By subscribing, you agree to our Privacy Policy and consent to receive updates from us.
               </p>
             </div>
           </div>
