@@ -1,6 +1,6 @@
 const config = {
-    plugins: {
-      "@tailwindcss/postcss": {},
-    },
-  };
-  export default config;
+  plugins: {
+    '@tailwindcss/postcss': {},
+  },
+};
+export default config;
